requirements.txt
icecream 
numpy
matplotlib
requests
flask
flask_cors
pillow
telepot